+++ libvast/vast/schema.hpp
@@ -13,6 +13,7 @@
 
 #pragma once
 
+#include "vast/defaults.hpp"
 #include "vast/detail/operators.hpp"
 #include "vast/detail/stable_set.hpp"
 #include "vast/type.hpp"
@@ -101,18 +102,21 @@ get_schema_dirs(const caf::actor_system_config& cfg,
                 std::vector<const void*> objpath_addresses = {nullptr});
 
 /// Loads a single schema file.
-/// @param sf The file path.
+/// @param schema_file The file path.
 /// @returns The parsed schema.
-caf::expected<schema> load_schema(const path& sf);
+caf::expected<schema> load_schema(const path& schema_file);
 
 /// Loads *.schema files from the given directories.
 /// @param schema_dirs The directories to load schemas from.
+/// @param max_recursion The maximum number of nested directories to traverse
+/// before aborting.
 /// @note Schemas from the same directory are merged, but directories are
 /// combined. It is designed so types that exist in later paths can override the
 /// earlier ones, but the same mechanism makes no sense inside of a single
 /// directory unless we specify a specific order of traversal.
 caf::expected<vast::schema>
-load_schema(const detail::stable_set<path>& schema_dirs);
+load_schema(const detail::stable_set<path>& schema_dirs,
+            size_t max_recursion = defaults::max_recursion);
 
 /// Loads schemas according to the configuration. This is a convenience wrapper
 /// around *get_schema_dirs* and *load_schema*.